requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-aiplatform==1.36.0
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.18.0
pandas==2.0.3
umap-learn==0.5.4
python-dotenv==1.0.0
streamlit==1.28.0
google-generativeai==0.3.1
flask-cors==4.0.0 